requirements.txt
# Core deep learning framework
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0

# Data handling & utilities
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
Pillow>=9.5.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Kaggle
kaggle>=1.6.6